LayoutTests/imported/w3c/web-platform-tests/mathml/presentation-markup/operators/display-operator-min-height-no-workaround-expected.txt


FAIL DisplayOperatorMinHeight is not modified assert_approx_equals: Using the small variant expected 40 +/- 3.2 but got 64
⫿
⫿
